preblocks: ERA5Normalizer, Gen2 trainer, smoke configs, test fixes
