+++ credit/preblock/scaler.py
@@ -23,7 +23,6 @@ class BridgeScalerTransformer(BasePreblock):
     """
 
     def __init__(self, scaler_path: str, variables: list[str], method: str):
-
         super().__init__()
         self.variables = variables
         self.method = method
@@ -31,5 +30,4 @@ class BridgeScalerTransformer(BasePreblock):
         self.scaler = load_scaler_dict(scaler_path)
 
     def forward(self, batch: dict) -> dict:
-
         return scale_var_dict(batch, self.scaler, self.method, self.variables)
